move port configuration into application package
